pyproject.toml
[project]
name = "niche-content-dashboard"
version = "0.1.0"
description = "Vision Model Evaluation Dashboard — interactive showcase of niche classification across 10 vision models"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "pillow>=10.0.0",
    "gunicorn>=22.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dashboard"]
